move getPassiveVoice specs to a folder
