+++ packages/yoastseo/spec/languageProcessing/researches/getPassiveVoice/getPassiveVoiceHungarianSpec.js
@@ -1,6 +1,6 @@
-import passiveVoice from "../../../src/languageProcessing/researches/getPassiveVoice.js";
-import Paper from "../../../src/values/Paper.js";
-import Researcher from "../../../src/languageProcessing/languages/hu/Researcher";
+import passiveVoice from "../../../../src/languageProcessing/researches/getPassiveVoice.js";
+import Paper from "../../../../src/values/Paper.js";
+import Researcher from "../../../../src/languageProcessing/languages/hu/Researcher";
 
 
 describe( "a test for detecting passive voice in sentences", function() {
