+++ packages/yoastseo/spec/languageProcessing/researches/getPassiveVoice/getPassiveVoicePolishSpec.js
@@ -1,6 +1,6 @@
-import passiveVoice from "../../../src/languageProcessing/researches/getPassiveVoice.js";
-import Paper from "../../../src/values/Paper.js";
-import Researcher from "../../../src/languageProcessing/languages/pl/Researcher";
+import passiveVoice from "../../../../src/languageProcessing/researches/getPassiveVoice.js";
+import Paper from "../../../../src/values/Paper.js";
+import Researcher from "../../../../src/languageProcessing/languages/pl/Researcher";
 
 describe( "detecting passive voice in sentences", function() {
 	it( "returns active voice (present)", function() {
