+++ packages/yoastseo/spec/languageProcessing/researches/getPassiveVoice/getPassiveVoiceSpanishSpec.js
@@ -1,6 +1,6 @@
-import passiveVoice from "../../../src/languageProcessing/researches/getPassiveVoice.js";
-import Paper from "../../../src/values/Paper.js";
-import Researcher from "../../../src/languageProcessing/languages/es/Researcher";
+import passiveVoice from "../../../../src/languageProcessing/researches/getPassiveVoice.js";
+import Paper from "../../../../src/values/Paper.js";
+import Researcher from "../../../../src/languageProcessing/languages/es/Researcher";
 
 describe( "detecting passive voice in sentences", function() {
 	it( "returns active voice (presente)", function() {
